Refactor seccomp and rdtsc code into seperate units
